+++ components/asset_util/common/fs.h
@@ -8,7 +8,8 @@
 const char* FS_Cwd(void);
 
 bool FS_FileExists(const char* qpath);
-size_t FS_FileSize(const char* qpath);
+//size_t FS_FileSize(const char* qpath);
+long int FS_FileSize(const char* qpath);
 
 const char* FS_GetExtensionSubString(const char* filename);
 void FS_StripExtension(const char* in, char* out);
